Update README.md
The ‘DEMONSTRATION OF COMPUTER GRAPHICS CONCEPTS USING MENU 
DRIVEN PROGRAM’ is a Computer Graphics and Visualization Lab Mini Project that 
illustrates the concepts in OpenGL. This project is based on a simple Menu Driven 
Program which demonstrates the CG concept. It has a menu list which provides six 
choices for the user by clicking the right button. Each choice provided will demonstrate 
a different concept in computer graphics. 
This Menu includes the following list: 
● Implement Bresenham’s line drawing algorithm 
● Clip a lines using Cohen-Sutherland algorithm 
● Draw a simple shaded scene consisting of a teapot on a table 
● Implement recursively subdivide a tetrahedron to form 3D sierpinski gasket 
● Implement a flag using Bezier Curve algorithm 
● Develop a program to fill the polygon using scan line algorithm 
All the 6 different programs of the Menu are made using openGL functions and 
libraries. This project reflects the use of the various openGL functions such as viewing, 
clipping, transformation, scaling, translation etc. The project involves mouse interaction 
by the user where the user is required to select a particular option from the menu list. 
The entire project environment is built using glut and the IDE used is code blocks 
developed by MinGW. 
The project provides an appealing user interface for the user with a plain background 
on which a particular output will be displayed on it based on the user choice. The user 
is required to click the right button on the mouse for the selection of the program. This 
project helps in understanding the few concepts of computer graphics practically with 
a different program for each.